replaced ushort by uint
Former-commit-id: d7534929bd1bb326ff054dec52aa671efcca7345

+++ test/Graphette/src/init/GenerateOrbitMatrix.cpp
@@ -21,7 +21,7 @@ void generateOrbitMatrix(uint n){
 	while(fin >> cgraph){
 		Graphette g = Graphette(n, cgraph);
 		vector<uint> idList(n);		
-		vector<vector<ushort>> orbits = g.getOrbits();
+		vector<vector<uint>> orbits = g.getOrbits();
 		for(auto orbit : orbits){
 			for(auto node : orbit){
 				idList[node] = orbitId;
